refactor: deleteCartProduct의 인자로 받는 id 타입을 Long에서 Int로 변경

+++ app/src/main/java/woowacourse/shopping/ui/cart/vm/CartViewModel.kt
@@ -25,7 +25,7 @@ class CartViewModel : ViewModel() {
         }
     }
 
-    fun deleteCartProduct(id: Long) {
+    fun deleteCartProduct(id: Int) {
         viewModelScope.launch {
             cartRepository.deleteCartProduct(id)
         }
